refactor: extract SectionTracker, optimize CSS loading, safe global usage
